refactor(tooltip): tooltip component refactor
