+++ packages/yike-design-ui/src/components/tooltip/index.ts
@@ -4,3 +4,4 @@ import { withInstall } from '@yike/utils';
 export const YkTooltip = withInstall(Tooltip);
 export default YkTooltip;
 export * from './src/tooltip';
+export * from './src/hooks';
